+++ mc2/controllers/docker/models.py
@@ -8,9 +8,12 @@ class DockerController(Controller):
     marathon_health_check_path = models.CharField(
         max_length=255, blank=True, null=True)
     port = models.PositiveIntegerField(default=0)
-    domain_urls = models.URLField(max_length=8000, default="")
+    domain_urls = models.TextField(max_length=8000, default="")
+    volume_needed = models.BooleanField(default=False)
+    volume_path = models.CharField(max_length=255, blank=True, null=True)
 
     def get_marathon_app_data(self):
+        app_data = super(DockerController, self).get_marathon_app_data()
         docker_dict = {
             "image": self.docker_image,
             "forcePullImage": True,
@@ -22,6 +25,20 @@ class DockerController(Controller):
                 "portMappings": [{"containerPort": self.port, "hostPort": 0}]
             })
 
+        parameters_dict = []
+        if self.volume_needed:
+            parameters_dict.append({"key": "volume-driver", "value": "xylem"})
+            parameters_dict.append({
+                "key": "volume",
+                "value": "%(app_id)s_media:%(path)s" % {
+                    'app_id': self.app_id,
+                    'path':
+                        self.volume_path or
+                        settings.MARATHON_DEFAULT_VOLUME_PATH}})
+
+        if parameters_dict:
+            docker_dict.update({"parameters": parameters_dict})
+
         domains = "%(generic_domain)s %(custom)s" % {
             'generic_domain': self.get_generic_domain(),
             'custom': self.domain_urls
@@ -29,20 +46,16 @@ class DockerController(Controller):
 
         service_labels = {
             "domain": domains.strip(),
+            "name": self.name,
         }
 
-        app_data = {
-            "id": self.app_id,
-            "cmd": self.marathon_cmd,
-            "cpus": self.marathon_cpus,
-            "mem": self.marathon_mem,
-            "instances": self.marathon_instances,
+        app_data.update({
             "labels": service_labels,
             "container": {
                 "type": "DOCKER",
                 "docker": docker_dict
             }
-        }
+        })
 
         if self.marathon_health_check_path:
             app_data.update({