+++ mc2/controllers/docker/tests/test_docker_controller.py
@@ -36,7 +36,8 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": "{}.{} {}".format(controller.app_id,
                                             settings.HUB_DOMAIN,
-                                            custom_urls)
+                                            custom_urls),
+                "name": "Test App",
             },
             "container": {
                 "type": "DOCKER",
@@ -60,7 +61,8 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": "{}.{} {}".format(controller.app_id,
                                             settings.HUB_DOMAIN,
-                                            custom_urls)
+                                            custom_urls),
+                "name": "Test App"
             },
             "container": {
                 "type": "DOCKER",
@@ -85,7 +87,8 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": "{}.{} {}".format(controller.app_id,
                                             settings.HUB_DOMAIN,
-                                            custom_urls)
+                                            custom_urls),
+                "name": "Test App",
             },
             "container": {
                 "type": "DOCKER",
@@ -108,6 +111,126 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             }]
         })
 
+        controller.volume_needed = True
+        controller.volume_path = "/deploy/media/"
+        controller.save()
+
+        self.assertEquals(controller.get_marathon_app_data(), {
+            "id": controller.app_id,
+            "cpus": 0.1,
+            "mem": 128.0,
+            "instances": 1,
+            "cmd": "ping",
+            "labels": {
+                "domain": "{}.{} {}".format(controller.app_id,
+                                            settings.HUB_DOMAIN,
+                                            custom_urls),
+                "name": "Test App",
+            },
+            "container": {
+                "type": "DOCKER",
+                "docker": {
+                    "image": "docker/image",
+                    "forcePullImage": True,
+                    "network": "BRIDGE",
+                    "portMappings": [{"containerPort": 1234, "hostPort": 0}],
+                    "parameters": [
+                        {"key": "volume-driver", "value": "xylem"},
+                        {
+                            "key": "volume",
+                            "value":
+                                "%s_media:/deploy/media/" % controller.app_id
+                        }]
+                }
+            },
+            "ports": [0],
+            "healthChecks": [{
+                "gracePeriodSeconds": 3,
+                "intervalSeconds": 10,
+                "maxConsecutiveFailures": 3,
+                "path": '/health/path/',
+                "portIndex": 0,
+                "protocol": "HTTP",
+                "timeoutSeconds": 5
+            }]
+        })
+
+        controller.volume_path = ""
+        controller.save()
+
+        self.assertEquals(controller.get_marathon_app_data(), {
+            "id": controller.app_id,
+            "cpus": 0.1,
+            "mem": 128.0,
+            "instances": 1,
+            "cmd": "ping",
+            "labels": {
+                "domain": "{}.{} {}".format(controller.app_id,
+                                            settings.HUB_DOMAIN,
+                                            custom_urls),
+                "name": "Test App",
+            },
+            "container": {
+                "type": "DOCKER",
+                "docker": {
+                    "image": "docker/image",
+                    "forcePullImage": True,
+                    "network": "BRIDGE",
+                    "portMappings": [{"containerPort": 1234, "hostPort": 0}],
+                    "parameters": [
+                        {"key": "volume-driver", "value": "xylem"},
+                        {
+                            "key": "volume",
+                            "value":
+                                "%s_media:%s" % (
+                                    controller.app_id,
+                                    settings.MARATHON_DEFAULT_VOLUME_PATH)
+                        }]
+                }
+            },
+            "ports": [0],
+            "healthChecks": [{
+                "gracePeriodSeconds": 3,
+                "intervalSeconds": 10,
+                "maxConsecutiveFailures": 3,
+                "path": '/health/path/',
+                "portIndex": 0,
+                "protocol": "HTTP",
+                "timeoutSeconds": 5
+            }]
+        })
+
+    def test_get_marathon_app_data_with_env(self):
+        controller = DockerController.objects.create(
+            name='Test App',
+            owner=self.user,
+            marathon_cmd='ping',
+            docker_image='docker/image',
+        )
+        self.mk_env_variable(controller)
+
+        self.assertEquals(controller.get_marathon_app_data(), {
+            "id": controller.app_id,
+            "cpus": 0.1,
+            "mem": 128.0,
+            "instances": 1,
+            "cmd": "ping",
+            "env": {"TEST_KEY": "a test value"},
+            "labels": {
+                "domain": "{}.{}".format(controller.app_id,
+                                         settings.HUB_DOMAIN),
+                "name": "Test App",
+            },
+            "container": {
+                "type": "DOCKER",
+                "docker": {
+                    "image": "docker/image",
+                    "forcePullImage": True,
+                    "network": "BRIDGE",
+                }
+            }
+        })
+
     @responses.activate
     def test_to_dict(self):
         controller = DockerController.objects.create(
@@ -127,4 +250,33 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             'marathon_cmd': 'ping',
             'port': 1234,
             'marathon_health_check_path': '/health/path/',
+        })
+
+    @responses.activate
+    def test_marathon_cmd_optional(self):
+        controller = DockerController.objects.create(
+            name='Test App',
+            owner=self.user,
+            docker_image='docker/image',
+        )
+
+        self.assertEquals(controller.get_marathon_app_data(), {
+            "id": controller.app_id,
+            "cpus": 0.1,
+            "mem": 128.0,
+            "instances": 1,
+            "cmd": "",
+            "labels": {
+                "domain": "{}.{}".format(controller.app_id,
+                                         settings.HUB_DOMAIN),
+                "name": "Test App",
+            },
+            "container": {
+                "type": "DOCKER",
+                "docker": {
+                    "image": "docker/image",
+                    "forcePullImage": True,
+                    "network": "BRIDGE",
+                }
+            }
         })